+++ src/MQTTPacketOut.h
@@ -25,9 +25,9 @@ void* MQTTPacket_connack(unsigned char aHeader, char* data, int datalen);
 
 int MQTTPacket_send_pingreq(networkHandles* net, char* clientID);
 
-int MQTTPacket_send_subscribe(List* topics, List* qoss, int msgid, int dup, networkHandles* net, char* clientID);
+int MQTTPacket_send_subscribe(List* topics, List* qoss, uint64_t msgid, int dup, networkHandles* net, char* clientID);
 void* MQTTPacket_suback(unsigned char aHeader, char* data, int datalen);
 
-int MQTTPacket_send_unsubscribe(List* topics, int msgid, int dup, networkHandles* net, char* clientID);
+int MQTTPacket_send_unsubscribe(List* topics, uint64_t msgid, int dup, networkHandles* net, char* clientID);
 
 #endif
